Fixed qmult again
Numpy's np.multiply behaves differently for and ndarray versus a matrix.
Taking a slice of a matrix and multiplying two columns will produce the
dot product of the two columns, as expected.  Doing the same on an
ndarray will give an extra dimension (2 identical columns).  This has
been fixed so no index out of bounds errors will occur in qmult